4d bin bigger input parameters

+++ include/global.hh
@@ -14,8 +14,8 @@
 typedef unsigned char uchar;        //!< The type to use for accessing individual bytes
 typedef int64_t llint;              //!< A large signed integer type.
 typedef float trace_t;              //!< Real type for traces, use "*_t" convention
-typedef double geom_t;              //!< Real type for geometry
-typedef double unit_t;              //!< Real type for geometry
+typedef float geom_t;               //!< Real type for geometry
+typedef float unit_t;               //!< Real type for geometry
 typedef const std::size_t csize_t;  //!< Short hand for a const size_t
 
 #include "anc/log.hh"
